fix(mesh): fix sonar issues

+++ engine/src/assets/Assets/Texture/TextureImporter.cpp
@@ -41,7 +41,7 @@ namespace nexo::assets {
             rendererTexture = renderer::NxTexture2D::create(std::get<ImporterFileInput>(ctx.input).filePath.string());
         else {
             auto data = std::get<ImporterMemoryInput>(ctx.input).memoryData;
-            rendererTexture = renderer::NxTexture2D::create(data.data(), data.size());
+            rendererTexture = renderer::NxTexture2D::create(data.data(), static_cast<unsigned int>(data.size()));
         }
         auto assetData = std::make_unique<TextureData>();
         assetData->texture = rendererTexture;
